Introduce PathBuilder/Storage abstractions
PathBuilder is intended to be a strict solution for stopping path
traversal vulnerabilities. There are two components to this, first, the
idea of a "root" that you can't escape out of, which is inspired by
golang <https://go.dev/blog/osroot>. The second is that each path
component has to be passed individually, and will be checked to ensure
there are no path traversal attempts.

Storage is a wrapper on top of PathBuilder, which implements some path
building functions we need in multiple places.

Finally, UnsafePathComponent is a way to indicate that a path component
is known to be potentially attacker-controlled, and typescript will
ensure that it is only ever used with PathBuilder and not e.g. the
standard library's path.join() or treated as a string.

Fixes #2771.