Add custom rule to prevent SQL injection
Three custom rules for the app:

1. Ban usage of db.exec() in favor of prepared statements
2. Don't allow dynamic content into prepared statements, with exceptions
3. Restrict creation of prepared statements to the database file.

This should codify our current practice and make it harder for any SQLi
attack to slip in.

selectWhereIn() is a deviation in that it dynamically creates a prepared
statement, so make it safer by requiring the table/column name to be explicitly
hardcoded.

The rules themselves were authored by Claude Code.