Use zod for parsing server responses
Introduce a clear parsing layer that can be used for validation
of server responses instead of blindly trusting them. As a starting
point, UUIDs are now validated across the board, which fixes a potential
path traversal since we use the UUIDs in file paths. Also validate
sources' journalist designation, since the client did so (because it
used it in file paths).

Add missing is_seen/has_attachment fields to Source that we were using
but never typed.

Fixes #2786.