Guard against path traversals in the gzip content's original filename
A malicious server could serve us a gzip file that contains a filename
that tries to exploit a path traversal attack. We do check this later
in the stack before further processing of the file, but it would let
an attacker (again, who compromised a server), to create a 0-byte file
wherever they'd like. Because the file will always be empty, at best this
can be used for a DoS attack and not code execution.

But it's simple enough to close this by adding an explicit check as we
parse the gzip file, so do so.

This was originally reported by AbhijitDas-Sukuna0007Abhi in BugCrowd;
thank you!

Fixes <https://github.com/freedomofpress/securedrop-security/issues/119>.