Memoize more React components
Most of these components just needed to be wrapped in memo(). In the
File case, semgrep thought the
InitialFile/InProgressFile/CompleteFile/FailedFile functions looked like
components but were really just functions that returned JSX. It's
simpler and logical to just treat them as real components, so converted
them to do that.

Memoization tests are added for all the components modified here.

Nearly all of this was done by Claude Code.