Switch translator-screenshots to use "test" mode
It was previously using a production build, which ends up
requiring a lot of special casing since we're not running this in
a SDW environment.

Also set `--no-qubes` for when it's being run in a Qubes VM.

Because of this, it'll also build the Rust securedrop-proxy during
the setup process.