Handle 403 errors gracefully
Send the user to the login screen with an error message. This
matches the current Python client's behavior.

We add a new SyncStatus.FORBIDDEN response and internally now return
statuses along with data (instead of e.g. `Index | null` where null is
304) so higher level code can return the correct status. Because the
statuses are hardcoded in the type, using the wrong one is a type error.

Fixes #2787.