Have dependabot upgrade better-sqlite3, node-abi with electron
Because we download better-sqlite3 binaries based on the specific
electron ABI, the two may need to be upgraded together.

Our download-sqlite-binaries.py script uses the node-abi package
(maintained by Electron) to know which ABI is in use, which also needs
to be updated alongside electron. Add it as a top-level dev dependency
since we directly use it and because we want dependabot to explicitly
update it.