Implement conffile override hack for all packages
Normally dpkg treats files in /etc/ as "conffiles", in which the user
can modify them and if so, their modifications will be preserved and
future updates from the package won't be applied.

But given our managed system, we don't want that to happen and all of
our /etc/ files should always be overridden with our new versions.
Instead of hardcoding specific packages, let's just apply it everywhere,
and it'll be a no-op if that package doesn't ship any /etc files.

Via lintian, this exposed that we were shipping files in /etc/ in
-client, -log and -proxy without the conffiles hack applying.

Fixes #1979.